v1.750 MythicMobs 5.2.0+ compatibility only (and still MC 1.19.2 only)

+++ src/main/java/com/gmail/berndivader/mythicmobsext/conditions/GetBowtensionCondition.java
@@ -35,7 +35,7 @@ public class GetBowtensionCondition extends AbstractCustomCondition implements I
 			} else {
 				if (p.getInventory().getItemInMainHand().getType() == Material.BOW
 						|| p.getInventory().getItemInOffHand().getType() == Material.BOW) {
-					d1 = (double) Utils.getBowTension((Player) e.getBukkitEntity());
+					d1 = Utils.getBowTension((Player) e.getBukkitEntity());
 				}
 			}
 		}
