+++ src/main/java/com/gmail/berndivader/mythicmobsext/utils/Vector.java
@@ -55,9 +55,9 @@ public class Vector implements Comparable<Vector> {
 	 * @param z the Z coordinate
 	 */
 	public Vector(int x, int y, int z) {
-		this.x = (double) x;
-		this.y = (double) y;
-		this.z = (double) z;
+		this.x = x;
+		this.y = y;
+		this.z = z;
 	}
 
 	/**
@@ -68,9 +68,9 @@ public class Vector implements Comparable<Vector> {
 	 * @param z the Z coordinate
 	 */
 	public Vector(float x, float y, float z) {
-		this.x = (double) x;
-		this.y = (double) y;
-		this.z = (double) z;
+		this.x = x;
+		this.y = y;
+		this.z = z;
 	}
 
 	/**
@@ -617,7 +617,7 @@ public class Vector implements Comparable<Vector> {
 	 * @param translateX what to add after rotation
 	 * @param translateZ what to add after rotation
 	 * @return a new vector
-	 * @see AffineTransform another method to transform vectors
+	 * @see com.gmail.berndivader.mythicmobsext.utils.math.transform.AffineTransform another method to transform vectors
 	 */
 	public Vector transform2D(double angle, double aboutX, double aboutZ, double translateX, double translateZ) {
 		angle = Math.toRadians(angle);
